+++ app/scripts/corpora/corpus-set.ts
@@ -8,6 +8,8 @@ import { objectIntersection, objectUnion } from "@/util"
 
 export type AttributeOption = Attribute & {
     group: "word" | "word_attr" | "sentence_attr"
+    /** Corpora that do not support this attribute */
+    unsupported: string[]
 }
 
 /** How to join attribute lists of different corpora */
@@ -51,10 +53,14 @@ export class CorpusSet {
         return this.corpora.map(f)
     }
 
+    getTokenCount(): number {
+        return sum(this.map((corpus) => parseInt(corpus.info.Size || "0")))
+    }
+
     getAttributes(lang?: string) {
         // lang not used here, only in parallel mode
         const attrs = this.map((corpus) => corpus.attributes)
-        return this._invalidateAttrs(attrs)
+        return objectUnion(attrs)
     }
 
     getAttributesIntersection() {
@@ -86,7 +92,7 @@ export class CorpusSet {
             const posAttrs = pickBy(corpus.attributes, (val, key) => val["is_struct_attr"])
             return { ...posAttrs, ...corpus["struct_attributes"] }
         })
-        const rest = this._invalidateAttrs(attrs)
+        const rest = objectUnion(attrs)
 
         // Merge datasets from attributes with the same name across all corpora
         for (const name in rest) {
@@ -121,19 +127,6 @@ export class CorpusSet {
         // Collect filters common to all corpora
         const attrs = intersection(...this.map((corpus) => corpus["attribute_filters"] || []))
         return pick(this.structAttributes, ...attrs)
-    }
-
-    _invalidateAttrs(attrs: Record<string, Attribute>[]) {
-        const union = objectUnion(attrs)
-        const intersection = objectIntersection(attrs)
-
-        // Mark attributes as disabled if not common to all attribute sets.
-        Object.entries(union).forEach(([key, value]) => {
-            if (!intersection[key]) value["disabled"] = true
-            else delete value["disabled"]
-        })
-
-        return union
     }
 
     /** Whether the given corpus has all given attributes. */
@@ -264,10 +257,12 @@ export class CorpusSet {
         const allAttrs = setOperator === "union" ? this.getAttributes(lang) : this.getAttributesIntersection()
 
         const attrs: AttributeOption[] = []
-        for (let key in allAttrs) {
-            const obj = allAttrs[key]
-            if (obj["display_type"] !== "hidden") {
-                attrs.push({ group: "word_attr", ...obj })
+        for (const attr of Object.values(allAttrs)) {
+            if (attr["display_type"] !== "hidden") {
+                const unsupported = this.corpora
+                    .filter((corpus) => !corpus.attributes[attr.name])
+                    .map((corpus) => corpus.id)
+                attrs.push({ group: "word_attr", ...attr, unsupported })
             }
         }
 
@@ -286,10 +281,12 @@ export class CorpusSet {
 
         let sentAttrs: AttributeOption[] = []
         const object = { ...common, ...allAttrs }
-        for (let key in object) {
-            const obj = object[key]
-            if (obj["display_type"] !== "hidden") {
-                sentAttrs.push({ group: "sentence_attr", ...obj })
+        for (const attr of Object.values(object)) {
+            if (attr["display_type"] !== "hidden") {
+                const unsupported = this.corpora
+                    .filter((corpus) => !corpus.struct_attributes[attr.name])
+                    .map((corpus) => corpus.id)
+                sentAttrs.push({ group: "sentence_attr", ...attr, unsupported })
             }
         }
 
@@ -297,7 +294,12 @@ export class CorpusSet {
     }
 
     getAttributeGroups(wordOp: SetOperator, structOp: SetOperator, lang?: string): AttributeOption[] {
-        const wordOption: AttributeOption = { group: "word", name: "word", label: settings["word_label"] }
+        const wordOption: AttributeOption = {
+            group: "word",
+            name: "word",
+            label: settings["word_label"],
+            unsupported: [],
+        }
         const attrs = this.getWordAttributeGroups(wordOp, lang)
         const sentAttrs = this.getStructAttributeGroups(structOp, lang)
         const comparator = (a: Attribute, b: Attribute) => locObj(a.label).localeCompare(locObj(b.label), getLang())