+++ app/scripts/backend/attr-values.ts
@@ -1,6 +1,7 @@
 import memoize from "lodash/memoize"
 import { AttrValuesResponseDeep, AttrValuesResponseFlat, RecursiveRecord } from "./types/attr-values"
 import { korpRequest } from "./common"
+import { uniq } from "lodash"
 
 /** Find which unique values occur and count them. */
 export const countAttrValues: (
@@ -31,15 +32,19 @@ export const getAttrValues: (
     attr: string,
     /** Whether values should be split by "|" */
     split?: boolean,
+    ranked?: boolean,
 ) => Promise<string[]> = memoize(
-    async (corpora, attr, split) => {
+    async (corpora, attr, split, ranked) => {
         const data = (await korpRequest("attr_values", {
             corpus: corpora.join(","),
             attr: attr,
             per_corpus: false,
             split: split ? attr : undefined,
         })) as AttrValuesResponseFlat
-        return data.combined[attr]
+        const values = data.combined[attr]
+        // If attribute is ranked, strip ":<rank>" suffix
+        if (ranked) return uniq(values.map((value) => value.replace(/:.*$/, "")))
+        return values
     },
     // Memoize based on the values of all arguments
     (...args) => JSON.stringify(args),