+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Show all attributes in grouping dropdown [#513](https://github.com/spraakbanken/korp-frontend/issues/513)
+
 ## [9.14.4] – 2026-05-06
 
 ### Fixed