+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+### Added
+
+- Show all attributes in grouping dropdown [#513](https://github.com/spraakbanken/korp-frontend/issues/513)
+- Show corpus name in tooltip of statistics table column heading
+
+### Fixed
+
+- Name attributes listed under both word and text attributes [#220](https://github.com/spraakbanken/korp-frontend/issues/220)
+- Change to "starts with", "contains", "ends with" [#516](https://github.com/spraakbanken/korp-frontend/issues/516)
+- HTML-escaped attribute values in trend diagram table
+- Reuse relative/absolute choice from statistics table in distribution pie chart
+
 ## [9.14.4] – 2026-05-06
 
 ### Fixed