+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Name attributes listed under both word and text attributes [#220](https://github.com/spraakbanken/korp-frontend/issues/220)
 - Change to "starts with", "contains", "ends with" [#516](https://github.com/spraakbanken/korp-frontend/issues/516)
+- Strip rank suffixes in extended search autocomplete [#518](https://github.com/spraakbanken/korp-frontend/issues/518)
 - HTML-escaped attribute values in trend diagram table
 - Reuse relative/absolute choice from statistics table in distribution pie chart