+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [9.15.0] – 2026-05-12
+
 ### Added
 
 - Show all attributes in grouping dropdown [#513](https://github.com/spraakbanken/korp-frontend/issues/513)
@@ -663,6 +665,7 @@
 - Lots of bug fixes for the sidebar
 
 [unreleased]: https://github.com/spraakbanken/korp-frontend/compare/master...dev
+[9.15.0]: https://github.com/spraakbanken/korp-frontend/releases/tag/v9.15.0
 [9.14.4]: https://github.com/spraakbanken/korp-frontend/releases/tag/v9.14.4
 [9.14.3]: https://github.com/spraakbanken/korp-frontend/releases/tag/v9.14.3
 [9.14.2]: https://github.com/spraakbanken/korp-frontend/releases/tag/v9.14.2